+++ lib/collect.js
@@ -13,6 +13,7 @@
 
 var constants = require('./constants');
 var settings = require('./settings').get();
+var getCodeBlock = require('./code-block');
 
 var lines;
 
@@ -69,17 +70,15 @@ function getComment(line, cursor){
     var indent = line.indexOf(flag);
 
     // get comment body 
-    var comment = getIndentedBlock(cursor, indent);
+    var comment = getCodeBlock(cursor, indent, lines);
 
     // get code block
     var root = lines[comment.cursor];
     var code = [root];
-    code.push(getIndentedBlock(comment.cursor, indent).body);
-
-    var type = getType(root);
+    code.push(getCodeBlock(comment.cursor, indent, lines).body);
 
     return {
-      body: comment.body,
+      comment: comment.body,
       code: code.join('\n'),
       indent: indent,
       type: getType(root),
@@ -91,37 +90,6 @@ function getComment(line, cursor){
 }
 
 
-/**
- * Get code block at line indent
- * returns 
- *   - code block
- *   - line cursor
- */
-
-function getIndentedBlock(cursor, indent){
-  var i = cursor;
-  var block = [];
-
-  while(i++){
-
-    // end of file
-    if(typeof lines[i] === 'undefined'){
-      break;
-    }
-
-    // indent match breaks
-    if(lines[i].match(/^\s*/g)[0].length <= indent){
-      break;
-    }
-
-    block.push(lines[i]);
-  }
-
-  return {
-    body: block.join('\n'),
-    cursor: i
-  };
-}
 
 
 /**
