+++ lib/generate.js
@@ -10,9 +10,12 @@ var YAML = require('js-yaml');
 var objectAssign = require('object-assign');
 
 var collect = require('./collect');
-var output = require('./output');
-var settings = require('./settings').get();
+var render = require('./render');
 
+
+/**
+ * Generate
+ */
 
 function generate(src, settings){
 
@@ -35,12 +38,12 @@ function generate(src, settings){
       jade: comment
     };
 
-    var json = YAML.safeLoad(comment.body);
+    var json = YAML.safeLoad(comment.comment);
     doc = objectAssign(doc, json);
 
     // ---
     // generate example
-    doc.output = output(doc);
+    doc.output = render(doc);
 
 
     // ---
@@ -65,6 +68,11 @@ function generate(src, settings){
   return store;
 }
 
+
+/**
+ * Complete
+ */
+
 function complete(store, settings){
 
   // create output dir if it doesn't exist
@@ -77,6 +85,8 @@ function complete(store, settings){
   fs.writeFile(settings.output, JSON.stringify(store, null, 2), function(err){
     if(err) throw err;
     console.log('Jade doc complete');
+
+    settings.complete();
   });
 }
