+++ lib/main.js
@@ -3,10 +3,10 @@
 
 var glob = require('glob');
 
-var settings = require('./settings');
 
 
 function JadeDoc(options){
+  var settings = require('./settings');
   settings = settings.set(options);
   
   var concat = require('./concat');
