+++ lib/render.js
