+++ lib/settings.js
@@ -13,7 +13,8 @@ function set(options){
   settings = objectAssign({
     output: 'jade-doc.json',
     directory: '.',
-    keyword: '@jadedoc'
+    keyword: '@jadedoc',
+    complete: function(){}
   }, options);
   
   return settings;
