+++ test/expected/mixins.json
@@ -1,26 +1,24 @@
 {
   "mixin1": {
-    "meta": {
-      "file": "./test/fixtures/mixins.jade"
-    },
     "jade": {
-      "root": "mixin mixin1",
       "comment": "  name: mixin1",
-      "block": "mixin mixin1\n  | this is a mixin",
-      "type": "mixin"
+      "code": "mixin mixin1\n  | this is a mixin",
+      "indent": 0,
+      "type": "mixin",
+      "root": "mixin mixin1",
+      "file": "./test/fixtures/mixins.jade"
     },
     "name": "mixin1",
     "output": "this is a mixin"
   },
   "mixin2": {
-    "meta": {
-      "file": "./test/fixtures/mixins.jade"
-    },
     "jade": {
-      "root": "mixin mixin2(arg1, arg2)",
       "comment": "  name: mixin2\n  description: this is jade mixin 2 documentation\n  arguments: \n    arg1: foo\n    arg2: faa",
-      "block": "mixin mixin2(arg1, arg2)\n  div this is a mixin #{arg1} #{arg2}",
-      "type": "mixin"
+      "code": "mixin mixin2(arg1, arg2)\n  div this is a mixin #{arg1}\n  div this is the same mixin #{arg2}",
+      "indent": 0,
+      "type": "mixin",
+      "root": "mixin mixin2(arg1, arg2)",
+      "file": "./test/fixtures/mixins.jade"
     },
     "name": "mixin2",
     "description": "this is jade mixin 2 documentation",
@@ -28,6 +26,18 @@
       "arg1": "foo",
       "arg2": "faa"
     },
-    "output": "<div>this is a mixin foo faa</div>"
+    "output": "<div>this is a mixin foo</div><div>this is the same mixin faa</div>"
+  },
+  "mixin3": {
+    "jade": {
+      "comment": "  name: mixin3",
+      "code": "//- mixin mixin3\n",
+      "indent": 0,
+      "type": "mixin",
+      "root": "//- mixin mixin3",
+      "file": "./test/fixtures/mixins.jade"
+    },
+    "name": "mixin3",
+    "output": "\n+//- mixin3"
   }
 }
